Refactor utils into multiple modules.
