+++ src/utils/mod.rs
@@ -1,0 +1,7 @@
+//! Various useful things.
+
+pub mod http;
+pub mod json;
+pub mod str;
+pub mod uri;
+
